+++ internal/provider/resource_meraki_device_management_interface.go
@@ -112,6 +112,10 @@ func (r *DeviceManagementInterfaceResource) Schema(ctx context.Context, req reso
 				ElementType:         types.StringType,
 				Optional:            true,
 			},
+			"wan1_vrf_name": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("The name of the VRF associated with WAN 1. If not provided, the default VRF is used.").String,
+				Optional:            true,
+			},
 			"wan2_static_gateway_ip": schema.StringAttribute{
 				MarkdownDescription: helpers.NewAttributeDescription("The IP of the gateway on the WAN.").String,
 				Optional:            true,
@@ -143,6 +147,31 @@ func (r *DeviceManagementInterfaceResource) Schema(ctx context.Context, req reso
 				MarkdownDescription: helpers.NewAttributeDescription("Up to two DNS IPs.").String,
 				ElementType:         types.StringType,
 				Optional:            true,
+			},
+			"wan2_vrf_name": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("The name of the VRF associated with WAN 2. If not provided, the default VRF is used.").String,
+				Optional:            true,
+			},
+			"ddns_hostnames_active_ddns_hostname": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Active dynamic DNS hostname.").String,
+				Computed:            true,
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.UseStateForUnknown(),
+				},
+			},
+			"ddns_hostnames_ddns_hostname_wan1": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("WAN 1 dynamic DNS hostname.").String,
+				Computed:            true,
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.UseStateForUnknown(),
+				},
+			},
+			"ddns_hostnames_ddns_hostname_wan2": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("WAN 2 dynamic DNS hostname.").String,
+				Computed:            true,
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.UseStateForUnknown(),
+				},
 			},
 		},
 	}