+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 - Fix "Missing Resource Identity After Read" provider error when a resource was first created with Terraform < 1.12 and has since been deleted out-of-band
 - Fix issue with `stack_ids` attribute of `meraki_network_vlan_profile_assignment` resource, [link](https://github.com/CiscoDevNet/terraform-provider-meraki/issues/224)
+- Add `wan1_vrf_name` and `wan2_vrf_name` attributes to `meraki_device_management_interface` resource and data source
+- Add `active_ddns_hostname`, `ddns_hostname_wan1`, and `ddns_hostname_wan2` computed attributes to `meraki_device_management_interface` resource and data source
 - Fix idempotency issue with `vpn_traffic_uplink_preferences[].traffic_filters`, `wan_traffic_uplink_preferences[].traffic_filters` attributes of `meraki_appliance_traffic_shaping_uplink_selection` resource, [link](https://github.com/CiscoDevNet/terraform-provider-meraki/pull/235)
 
 ## 1.12.1